Modify measurementUnits to include dual-type units and custom units
Specifically, tsp, tbsp, oz, and cups can be used for dry food as well
with a slight loss of precision. A custom 'serving(s)' type was
introduced for situations in which it would be impractical to count
by weight or volume (such as a loaf of bread).

Imperial units have been commented out for the time being until their
use case becomes apparent.